// src/lib/types/entity/action-history-record.ts
/*
 *  AtroCore Software
 *
 *  This source file is available under GNU General Public License version 3 (GPLv3).
 *  Full copyright and license information is available in LICENSE.txt, located in the root directory.
 *
 *  @copyright  Copyright (c) AtroCore GmbH (https://www.atrocore.com)
 *  @license    GPLv3 (https://www.gnu.org/licenses/)
 */

type ActionHistoryRecord = {
    controllerName: string;
    targetId: string | null;
    targetName: string | null;
    targetUrl: string | null;
};

export default ActionHistoryRecord;